+++ src/components/property-page-widgets/PropertyPageHero/component.js
@@ -32,6 +32,7 @@ const Component = ({
   searchBarModalHeadingText,
   searchBarSearchButton,
   secondaryButtonText,
+  searchBarOnChangeInput,
   searchBarOnSubmit,
 }) => (
   <Hero
@@ -47,6 +48,7 @@ const Component = ({
     headerSearchBarSearchButton={searchBarSearchButton}
     placeholderBackgroundImageUrl={images[0].placeholderImageUrl}
     searchBarGetIsDayBlocked={searchBarGetIsDayBlocked}
+    searchBarOnChangeInput={searchBarOnChangeInput}
     searchBarOnSubmit={searchBarOnSubmit}
   >
     <FlexContainer alignItems="flex-end">
@@ -83,6 +85,7 @@ Component.defaultProps = {
   searchBarModalHeadingText: undefined,
   searchBarSearchButton: undefined,
   secondaryButtonText: VIEW_MORE_PICTURES,
+  searchBarOnChangeInput: undefined,
   searchBarOnSubmit: undefined,
 };
 
@@ -181,6 +184,14 @@ Component.propTypes = {
   ).isRequired,
   /* The heading displayed in the search bar modal. */
   searchBarModalHeadingText: PropTypes.string,
+  /** A function called when a change in an input occurs in the search bar.
+   *  @param {Object} values - The values of the inputs in the search bar.
+   *  @param {Object} values.dates
+   *  @param {String} values.guests
+   *  @param {String} values.location
+   */
+  // eslint-disable-next-line react/no-unused-prop-types
+  searchBarOnChangeInput: PropTypes.func,
   /** The function to call when the search bar is submitted.
    *  @param {Object} values - The values of the inputs in the search bar.
    *  @param {Object} values.dates